Update README.md
This project is an Ionic Angular application for managing a phone book. Users can view, add, edit, and search contacts. The app features a responsive UI with forms for entering contact details (name, phone number, and optional email), real-time search filtering, and integration with a backend API for persistent storage